+++ src/store/modules/setting/type.ts
@@ -1,6 +1,21 @@
 import type { GlobalThemeOverrides } from 'naive-ui'
 import type { Placement } from '@/types/modules/component'
-import type { APP_WATERMARK_CONFIG } from '@/app-config/appConfig'
+
+export interface WatermarkConfig {
+  content: string
+  fontSize: number
+  lineHeight: number
+  width: number
+  height: number
+  xOffset: number
+  yOffset: number
+  rotate: number
+}
+
+export interface AppRootRoute {
+  name: string
+  path: string
+}
 
 export interface SettingState {
   drawerPlacement: Placement
@@ -13,5 +28,6 @@ export interface SettingState {
   watermarkSwitch: boolean
   copyrightSwitch: boolean
   contentTransition: string
-  watermarkConfig: typeof APP_WATERMARK_CONFIG
+  watermarkConfig: WatermarkConfig
+  appRootRoute: AppRootRoute
 }